% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/runPCA.r
\name{choosePC}
\alias{choosePC}
\title{choosePC determines the N of PCs to consider to account for ~50% fo the variability
the function finds cumulative variance of Principal Components and find the number of PCs
required to account for 50% of the variance}
\usage{
choosePC(pcaObj, varThreshold = 0.5)
}
\arguments{
\item{pcaObj}{the PCA object}

\item{variance}{the threshold amt of variance that is required}
}
\description{
choosePC determines the N of PCs to consider to account for ~50% fo the variability
the function finds cumulative variance of Principal Components and find the number of PCs
required to account for 50% of the variance
}

