% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/runPCA.r
\name{closest}
\alias{closest}
\title{closest finds videos to recommend based on the PCA matrix
Closest finds the closest users (Ux) for each user (U0)
selects videos which they (Uxs) have watched and recommends to U0
videos watched by Uxs but not U0}
\usage{
closest(latentVars, distance = "euclidian", top = 5, users, behaviour,
  cores = 1)
}
\arguments{
\item{latentVars}{could be as many N as possible}

\item{distance}{the type of distance formula to use, defaults to euclidian distance}

\item{users}{A list of userIDs in the same order as was used in the matrix to generate the PCA}

\item{integer}{representing the top N closest users; defaults to 5}
}
\description{
closest finds videos to recommend based on the PCA matrix
Closest finds the closest users (Ux) for each user (U0)
selects videos which they (Uxs) have watched and recommends to U0
videos watched by Uxs but not U0
}

