% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/runPCA.r
\name{runPCA}
\alias{runPCA}
\title{runPCA runs principal component analysis using the prcomp function
\code{runPCA} takes the first N PCs (latent variables) and
maps users onto the new N dimensional space}
\usage{
runPCA(df)
}
\arguments{
\item{df}{data.frame for genre table generated by genre2matrix in the tools folder}
}
\description{
runPCA runs principal component analysis using the prcomp function
\code{runPCA} takes the first N PCs (latent variables) and
maps users onto the new N dimensional space
}

