% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/submission.R
\name{submit}
\alias{submit}
\title{submit to join the predictions}
\usage{
submit(monthOne, monthTwo, filePath)
}
\arguments{
\item{monthOne}{}

\item{monthTwo}{}

\item{filePath}{}
}
\description{
\code{submit} concatenates the predictions for the
two months and generates the submission file as a csv
}

