% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/runPCA.r
\name{whichVideos}
\alias{whichVideos}
\title{whichVideos choses which videos watched by the user's neighbours to recommend
the description}
\usage{
whichVideos(candidates, behaviour)
}
\arguments{
\item{candidate}{}
}
\description{
whichVideos choses which videos watched by the user's neighbours to recommend
the description
}

